Se mueven entidades del paquete "entidades" al paquete "logica.entidades" para que queden en la "capa de logica" segun lo requerido por el profesor
