+++ src/main/java/logica/entidades/Cuenta.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entidades;
+package logica.entidades;
 
 import java.io.Serializable;
 import java.util.Date;
