+++ src/main/resources/META-INF/persistence.xml
@@ -2,12 +2,12 @@
 <persistence version="2.1" xmlns="http://xmlns.jcp.org/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence http://xmlns.jcp.org/xml/ns/persistence/persistence_2_1.xsd">
   <persistence-unit name="miBilleteraPU" transaction-type="RESOURCE_LOCAL">
     <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-    <class>entidades.Cuenta</class>
     <class>entidades.Categoria</class>
-    <class>entidades.Movimiento</class>
-    <class>entidades.CategoriaMovimiento</class>
-    <class>entidades.TipoCategoria</class>
-    <class>entidades.IngresosVsEgresos</class>
+    <class>logica.entidades.Movimiento</class>
+    <class>logica.entidades.CategoriaMovimiento</class>
+    <class>logica.entidades.TipoCategoria</class>
+    <class>logica.entidades.Cuenta</class>
+    <class>logica.entidades.IngresosVsEgresos</class>
     <properties>
       <property name="javax.persistence.jdbc.url" value="jdbc:sqlite::resource:miBilletera.db"/>
       <property name="javax.persistence.jdbc.user" value="root"/>
